+++ src/lib/eina/eina_sched.c
@@ -31,94 +31,8 @@
 #include "eina_sched.h"
 #include "eina_log.h"
 
-#define RTNICENESS 1
-#define NICENESS 5
-
 EAPI void
 eina_sched_prio_drop(void)
 {
-#ifndef _WIN32
-   struct sched_param param;
-   int pol, ret;
-   Eina_Thread pthread_id;
-
-   pthread_id = eina_thread_self();
-   ret = pthread_getschedparam(pthread_id, &pol, &param);
-   if (ret)
-   {
-      EINA_LOG_ERR("Unable to query sched parameters");
-      return;
-   }
-
-   if (EINA_UNLIKELY(pol == SCHED_RR || pol == SCHED_FIFO))
-   {
-      param.sched_priority -= RTNICENESS;
-
-      /* We don't change the policy */
-      if (param.sched_priority < 1)
-      {
-         EINA_LOG_INFO("RT prio < 1, setting to 1 instead");
-         param.sched_priority = 1;
-      }
-
-      pthread_setschedparam(pthread_id, pol, &param);
-   }
-#ifdef __linux__
-   else
-   {
-      int prio;
-      errno = 0;
-      prio = getpriority(PRIO_PROCESS, 0);
-      if (errno == 0)
-      {
-         prio += NICENESS;
-         if (prio > 19)
-         {
-            EINA_LOG_INFO("Max niceness reached; keeping max (19)");
-            prio = 19;
-         }
-
-         setpriority(PRIO_PROCESS, 0, prio);
-      }
-   }
-#endif
-#else
-   Eina_Thread pthread_id;
-   int sched_priority;
-
-   pthread_id = eina_thread_self();
-
-   sched_priority = GetThreadPriority((HANDLE)pthread_id);
-
-   if (EINA_UNLIKELY(sched_priority == THREAD_PRIORITY_TIME_CRITICAL))
-   {
-      sched_priority -= RTNICENESS;
-
-      /* We don't change the policy */
-      if (sched_priority < 1)
-      {
-         EINA_LOG_INFO("RT prio < 1, setting to 1 instead");
-         sched_priority = 1;
-      }
-      if (!SetThreadPriority((HANDLE)pthread_id, sched_priority))
-      {
-         EINA_LOG_ERR("Unable to query sched parameters");
-      }
-   }
-   else
-   {
-      sched_priority += NICENESS;
-
-      /* We don't change the policy */
-      if (sched_priority > THREAD_PRIORITY_TIME_CRITICAL)
-      {
-         EINA_LOG_INFO("Max niceness reached; keeping max (THREAD_PRIORITY_TIME_CRITICAL)");
-         sched_priority = THREAD_PRIORITY_TIME_CRITICAL;
-      }
-      if (!SetThreadPriority((HANDLE)pthread_id, sched_priority))
-      {
-         EINA_LOG_ERR("Unable to query sched parameters");
-      }
-   }
-#endif
+   _eina_sched_prio_drop();
 }
